added python file for model selection
Fifteen different ML classifiers are used to fit the data and the corresponding accuracy scores are printed. This function can be imported for any text data for finding the models with the best accuracy scores. Provided that the train and test data must be provided in the form of cleaned and tokenized text.